Move helper function to quantizers.py
